+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Unreleased
+- **added:** implemented `Default` for `Stats`, `ChunkPrevIter`, `ChunkNextIter`, `WithoutShrink` and `WithoutDealloc`
+
 ## 0.15.1 (2025-01-04)
 - **added:** implemented `TakeOwnedSlice` for `vec::IntoIter` and `vec::Drain`
 - **performance:** improved speed and time complexity of `split_off`